refactor: 카드리스트 애니메이션 로직 hook으로 분리 후 적용
