+++ src/hooks/index.ts
@@ -8,3 +8,4 @@ export { default as useIsOverflow } from './useIsOverflow';
 export { default as useToggle } from './useToggle';
 export { default as useScrollToBottom } from './useScrollToBottom';
 export { default as useInterval } from './useInterval';
+export { default as useScrollAnimation } from './useScrollAnimation';
